+++ packages/preact_signals/CHANGELOG.md
@@ -3,3 +3,4 @@
 - Adding `computed`, `effect`, `signal`, `untracked`
 - Adding tests
 - Adding web example
+- Adding extensions for `Future` and `Stream`